backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
easyocr
numpy
pillow
